Move project initializer to runs/
