+++ layer/main/run.py
@@ -14,9 +14,9 @@ from layer.global_context import (
     reset_to,
     set_has_shown_python_version_message,
 )
-from layer.projects.init_project_runner import InitProjectRunner
 from layer.projects.project_runner import ProjectRunner
 from layer.projects.utils import get_current_project_full_name, validate_project_name
+from layer.runs.initializer import InitProjectRunner
 from layer.settings import LayerSettings
 from layer.utils.async_utils import asyncio_run_in_thread
